First attempt at rendering the field. Initial logic for setting up the kicking team

+++ modules/game-ui/common/src/commonMain/kotlin/dk/ilios/jervis/ui/model/GameStatusViewModel.kt
@@ -10,7 +10,7 @@ class GameStatusViewModel(val controller: GameController) {
 
     fun progress(): Flow<GameProgress> {
         return controller.state.gameFlow.map { game ->
-            GameProgress(game.halfNo, game.driveNo, game.activeTeam.turnData.currentTurn, game.activeTeam.name)
+            GameProgress(game.halfNo.toInt(), game.driveNo, game.activeTeam.turnData.currentTurn.toInt(), game.activeTeam.name)
         }
 
     }
